% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{launch_server}
\alias{launch_server}
\title{Launch example server}
\usage{
launch_server(
  port = 4567,
  title = "ps200b: Example server",
  eg_str = "eg00_estimating_a_mean_10",
  ...
)
}
\description{
Launch example server
}
